added integration tests

+++ src/main/java/ru/hogwarts/schoolloohcs/services/FacultyServiceImpl.java
@@ -1,10 +1,12 @@
 package ru.hogwarts.schoolloohcs.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import ru.hogwarts.schoolloohcs.model.Faculty;
 import ru.hogwarts.schoolloohcs.repository.FacultyRepository;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -30,8 +32,11 @@ public class FacultyServiceImpl implements FacultyService {
 
     //Delete faculty
     @Override
-    public void deleteFaculty(long id){
-        facultyRepository.deleteById(id);
+    public Faculty deleteFaculty(long id){
+        Faculty fac = facultyRepository.findById(id).orElse(null);
+        if (fac != null){
+            facultyRepository.deleteById(id);
+        } return fac;
     }
 
     //Edit faculty
